# openapi.yaml
components:
  schemas:
...
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/v1/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    ListNotificationsOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/v1/schemas/ListNotificationsOutputBody.json
          format: uri
          readOnly: true
          type: string
        notifications:
          items:
            $ref: "#/components/schemas/NotificationBody"
          type:
            - array
            - "null"
        page:
          format: int64
          type: integer
        per_page:
          format: int64
          type: integer
        total:
          format: int64
          type: integer
      required:
        - notifications
        - total
        - page
        - per_page
      type: object
...
    NotificationBody:
      additionalProperties: false
      properties:
        content:
          type: string
        date:
          format: date-time
          type: string
        id:
          format: int64
          type: integer
        title:
          type: string
      required:
        - id
        - title
        - content
        - date
      type: object
...
info:
  title: flagfish
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /notifications:
    get:
      operationId: list-notifications
      parameters:
        - explode: false
          in: query
          name: page
          schema:
            default: 1
            format: int64
            maximum: 1000000
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: per_page
          schema:
            default: 50
            format: int64
            maximum: 100
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListNotificationsOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List notifications (paginated)
      tags:
        - notifications
...
servers:
  - url: /api/v1